pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seedance-aspect"
version = "0.1.0"
description = "Seedance 2.0 powered horizontal/vertical video aspect remake CLI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "tos>=2.7.0",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
seedance-aspect = "seedance_aspect.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/seedance_aspect"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

